+++ spec/features/staff/users_can_view_reports_spec.rb
@@ -78,10 +78,17 @@ RSpec.feature "Users can view reports" do
       expect(page).to have_content report.organisation.name
     end
 
-    scenario "the report inclides a list of updated activities" do
-      activity = create(:third_party_project_activity)
-      report = create(:report, :active, organisation: build(:delivery_partner_organisation))
-      history_event = create(:historical_event, activity: activity, report: report)
+    scenario "the report includes a list of newly created and updated activities" do
+      delivery_partner_organisation = create(:delivery_partner_organisation)
+      fund = create(:fund_activity)
+      programme = create(:programme_activity, parent: fund)
+      project = create(:project_activity, parent: programme)
+      report = create(:report, :active, fund: fund, organisation: delivery_partner_organisation, financial_quarter: 1, financial_year: 2021)
+
+      new_activity = create(:third_party_project_activity, organisation: delivery_partner_organisation, parent: project, originating_report: report)
+      updated_activity = create(:third_party_project_activity, organisation: delivery_partner_organisation, parent: project)
+
+      _history_event = create(:historical_event, activity: updated_activity, report: report)
 
       visit reports_path
 
@@ -93,7 +100,8 @@ RSpec.feature "Users can view reports" do
         click_on "Activities"
       end
 
-      expect(page).to have_content activity.title
+      expect(page).to have_content new_activity.title
+      expect(page).to have_content updated_activity.title
     end
 
     context "when there is no report descripiton" do